Add interview scene; Adjusts in protest view; Add fadein fadeout for all transition; Add speedup and skip for animation in interview and protest scene;

+++ Assets/Script/Ocupation/Ocupation.cs
@@ -24,7 +24,7 @@ namespace Assets.Script.Ocupation
         public int _cookStatus = 3;
         public int _startPersonQuantity = 0;
         public int _persons = 0;
-        public float _popularAdeptance = 0;
+        public int _popularAdeptance = 0;
 
         void Start()
         {
